# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShoeFlow Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whatsapp/sessions:
    get:
      operationId: getWhatsappSessions
      tags: [whatsapp]
      summary: Get all WhatsApp sessions
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhatsappSession"
    post:
      operationId: createWhatsappSession
      tags: [whatsapp]
      summary: Create a new WhatsApp session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappSessionInput"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappSession"
  /whatsapp/sessions/{id}:
    get:
      operationId: getWhatsappSession
      tags: [whatsapp]
      summary: Get a WhatsApp session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappSession"
    delete:
      operationId: deleteWhatsappSession
      tags: [whatsapp]
      summary: Delete a WhatsApp session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /whatsapp/sessions/{id}/qr:
    get:
      operationId: getSessionQr
      tags: [whatsapp]
      summary: Get QR code for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all WhatsApp groups
      parameters:
        - name: sessionId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Groups list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Manually register a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a single group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
...
  /messages/forwarded:
    get:
      operationId: listForwardedMessages
      tags: [messages]
      summary: List forwarded messages
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForwardedMessage"
    post:
      operationId: forwardMessage
      tags: [messages]
      summary: Forward a message to groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForwardInput"
      responses:
        "201":
          description: Forwarding queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForwardedMessage"
  /messages/responses:
    get:
      operationId: listResponses
      tags: [messages]
      summary: List detected responses and availability
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [disponible, agotado, pendiente, confirmado]
      responses:
        "200":
          description: Responses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShoeResponse"
  /messages/responses/{id}:
    patch:
      operationId: updateResponse
      tags: [messages]
      summary: Update a response status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShoeResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    WhatsappSession:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        numero:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        estado:
          type: string
          enum: [conectado, sincronizando, desconectado, reconectando]
        gruposSincronizados:
          type: integer
        creadoEn:
          type: string
        ultimaConexion:
          type: ["string", "null"]
      required: [id, nombre, estado, gruposSincronizados, creadoEn]
    WhatsappSessionInput:
      type: object
      properties:
        nombre:
          type: string
      required: [nombre]
    QrCode:
      type: object
      properties:
        qrData:
          type: ["string", "null"]
        status:
          type: string
        expiraEn:
          type: integer
      required: [status]
    Group:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        jid:
          type: ["string", "null"]
        nombre:
          type: string
        descripcion:
          type: ["string", "null"]
        foto:
          type: ["string", "null"]
        categoria:
          type: ["string", "null"]
        participantes:
          type: integer
        mensajesDiarios:
          type: integer
        ultimaActividad:
          type: ["string", "null"]
        activo:
          type: boolean
        creadoEn:
          type: string
      required: [id, sessionId, nombre, participantes, mensajesDiarios, activo, creadoEn]
    GroupInput:
      type: object
      properties:
        sessionId:
          type: integer
        nombre:
          type: string
        jid:
          type: string
        categoria:
          type: string
        participantes:
          type: integer
      required: [sessionId, nombre]
    GroupUpdate:
      type: object
      properties:
        nombre:
          type: string
        categoria:
          type: string
        descripcion:
          type: string
        activo:
          type: boolean
...
    ForwardedMessage:
      type: object
      properties:
        id:
          type: integer
        contenido:
          type: string
        grupoOrigen:
          type: ["string", "null"]
        gruposDestino:
          type: array
          items:
            type: string
        estado:
          type: string
          enum: [pendiente, enviando, completado, fallido]
        progreso:
          type: integer
        timestamp:
          type: string
        proveedor:
          type: ["string", "null"]
        referencia:
          type: ["string", "null"]
      required: [id, contenido, gruposDestino, estado, progreso, timestamp]
    ForwardInput:
      type: object
      properties:
        contenido:
          type: string
        grupoOrigenId:
          type: integer
        gruposDestinoIds:
          type: array
          items:
            type: integer
      required: [contenido, gruposDestinoIds]
    ShoeResponse:
      type: object
      properties:
        id:
          type: integer
        referencia:
          type: string
        imagen:
          type: ["string", "null"]
        proveedorNombre:
          type: string
        proveedorNumero:
          type: string
        precio:
          type: ["string", "null"]
        grupoOrigen:
          type: string
        estado:
          type: string
          enum: [disponible, agotado, pendiente, confirmado]
        timestamp:
          type: string
        prioridad:
          type: boolean
      required:
        - id
        - referencia
        - proveedorNombre
        - proveedorNumero
        - grupoOrigen
        - estado
        - timestamp
        - prioridad
    ResponseUpdate:
      type: object
      properties:
        estado:
          type: string
          enum: [disponible, agotado, pendiente, confirmado]
        prioridad:
          type: boolean
...